Factor out A2A item parser.
